+++ composer.json
@@ -11,6 +11,7 @@
     ],
     "require": {
         "php": ">=7.0",
-        "composer/installers": "~1.0"
+        "composer/installers": "~1.0",
+        "october/system": ">=2.1.12"
     }
 }
